requirements.txt
alembic==1.11.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
qrcode==7.4.2
Pillow==10.1.0
reportlab==4.0.7
openpyxl==3.1.5
python-dotenv==1.0.1
twilio==8.4.1
twilio>=8.0.0
# Added runtime dependency required by pynacl
cffi>=1.15.1

source venv/bin/activate
python app.py

curl -L https://github.com/cloudflare/cloudflared/releases/latest/download/cloudflared-linux-amd64.deb -o cloudflared.deb
sudo dpkg -i cloudflared.deb

cloudflared tunnel --url http://localhost:5000